backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
sqlmodel>=0.0.16,<1
psycopg2-binary>=2.9,<3
redis>=5.0,<6
pydantic-settings>=2.0,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9,<1
httpx>=0.27,<1
pillow==12.2.0